+++ test/block_dialog_show_test.dart
@@ -19,7 +19,9 @@ class _TestApp extends StatelessWidget {
                     BlockText(text: 'Hello'),
                   ]),
                   BlockRow(blocks: [
-                    BlockButton(label: 'Close', onPressed: (_) async {}),
+                    BlockButton(
+                        label: 'Close',
+                        onPressed: (result, controller) async {}),
                   ]),
                 ],
               );